Move pagination class to its own file
